:green_heart: issue #13 カスタムプログレスバーのテストコードを ViewInspector で記述
